[select] upd focus story and added test
